# backend/functions/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sparkling REST API
  version: 1.0.0
  description: >-
    Sparkling platform backend (SRS v1.0). All JSON, snake_case, ISO-8601 UTC timestamps, money in integer ZAR cents. Mutating routes accept `Idempotency-Key` (or `client_op_id` in the body); replays
    return the cached first response. Every response carries `X-Correlation-Id`. Clients send `X-Client-App` and `X-Client-Version`.
  contact:
    name: Sparkling engineering
    email: dev@sparkling.co.za
  license:
    name: Proprietary
    url: https://sparkling.co.za/legal
servers:
  - url: https://europe-west1-sparkling-4e89d.cloudfunctions.net/api/v1
    description: dev (Cloud Functions 2nd gen)
  - url: http://127.0.0.1:5001/sparkling-4e89d/europe-west1/api/v1
    description: Firebase emulator
...
security:
  - bearerAuth: []
paths:
...
  /vehicles:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Vehicles
      operationId: listVehicles
      summary: Own vehicles (customer) or search (staff)
      parameters:
        - name: customer_id
          in: query
          required: false
          schema:
            type: string
        - name: registration_no
          in: query
          required: false
          schema:
            type: string
    post:
      responses:
        '200':
          description: Duplicate replay / forced merge
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicle:
                    $ref: '#/components/schemas/Vehicle'
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicle:
                    $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Vehicles
      operationId: createVehicle
      summary: Add a vehicle (409 with existing_vehicle_id on duplicate unless force)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
...
  /staff/customers:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StaffCustomer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Staff
      operationId: searchStaffCustomers
      summary: Search customers by name, phone (any formatting), e-mail or plate (audited customer.search)
      parameters:
        - name: search
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 80
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 25
            default: 10
    post:
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    $ref: '#/components/schemas/StaffCustomer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Staff
      operationId: createStaffCustomer
      summary: Register a walk-in customer (walkin_<uuid> profile, claimed on sign-up by e-mail or phone); 409 with existing_customer when the phone/e-mail exists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffCustomerCreate'
  /staff/customers/{id}/vehicles:
    post:
      responses:
        '200':
          description: Duplicate replay / forced merge
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicle:
                    $ref: '#/components/schemas/Vehicle'
                  duplicate:
                    type: boolean
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicle:
                    $ref: '#/components/schemas/Vehicle'
                  duplicate:
                    type: boolean
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Staff
      operationId: createStaffCustomerVehicle
      summary: Add a vehicle for a customer (409 with existing_vehicle_id on duplicate plate/VIN unless force)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  responses:
    ValidationError:
      description: validation_error (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      description: unauthenticated (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: forbidden (403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: not_found (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: conflict / invalid_transition (409)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Gone:
      description: gone (410) — expired quote link or quotation past valid_until
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: rate_limited (429)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Internal:
      description: internal (500) — never includes stack traces
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - not_found
                - validation_error
                - conflict
                - invalid_transition
                - invalid_otp
                - gone
                - rate_limited
                - internal
            message:
              type: string
            details:
              type: array
              items: {}
            correlation_id:
              type: string
          required:
            - code
            - message
            - correlation_id
      required:
        - error
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
        registration_no:
          type: string
        vin:
          type: string
          nullable: true
        engine_no:
          type: string
          nullable: true
        make:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        colour:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        licence_no:
          type: string
          nullable: true
        disc_expiry:
          type: string
          format: date
          nullable: true
        source:
          type: string
          enum:
            - manual
            - scan
        disc_verified:
          type: boolean
        disc_hash:
          type: string
          nullable: true
        size_class:
          type: string
          enum:
            - small
            - large
            - bike
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    StaffCustomer:
      type: object
      properties:
        id:
          type: string
          description: Firebase UID, or walkin_<uuid> until claimed
        full_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        marketing_opt_in:
          type: boolean
        whatsapp_opt_in:
          type: boolean
        loyalty:
          type: object
          properties:
            tier:
              type: string
              enum:
                - silver
                - gold
                - platinum
                - black
            balance_points:
              type: integer
            discount_pct:
              type: number
            plan_code:
              type: string
              example: gold
              nullable: true
            plan_name:
              type: string
              nullable: true
            membership_status:
              type: string
              enum:
                - pending
                - active
                - past_due
                - cancelled
                - expired
              nullable: true
            included_remaining:
              type: integer
              description: Remaining monthly washes on the live plan (0 without one)
          required:
            - tier
            - balance_points
          nullable: true
        vehicles:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              registration_no:
                type: string
              make:
                type: string
                nullable: true
              model:
                type: string
                nullable: true
              colour:
                type: string
                nullable: true
              disc_verified:
                type: boolean
      required:
        - id
        - full_name
        - email
        - phone
        - marketing_opt_in
        - whatsapp_opt_in
        - loyalty
        - vehicles
    StaffCustomerCreate:
      type: object
      properties:
        full_name:
          type: string
          minLength: 2
        phone:
          type: string
          description: Any SA/international format; stored as E.164 (082… → +2782…)
        email:
          type: string
          format: email
          nullable: true
        marketing_opt_in:
          type: boolean
        whatsapp_opt_in:
          type: boolean
        outlet_id:
          type: string
          format: uuid
        client_op_id:
          type: string
      required:
        - full_name
        - phone
        - client_op_id
...
    VehicleCreate:
      type: object
      properties:
        registration_no:
          type: string
        vin:
          type: string
          minLength: 17
          maxLength: 17
          nullable: true
        make:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        colour:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        licence_no:
          type: string
          nullable: true
        disc_expiry:
          type: string
          format: date
          nullable: true
        engine_no:
          type: string
          nullable: true
        source:
          type: string
          enum:
            - manual
            - scan
        disc_hash:
          type: string
          description: sha256 of the raw disc payload (raw is never stored)
          nullable: true
        force:
          type: boolean
        client_op_id:
          type: string
        size_class:
          type: string
          enum:
            - small
            - large
            - bike
          nullable: true
        description:
          type: string
          description: Licence-disc description; only used to derive size_class when it is omitted
          nullable: true
      required:
        - registration_no
...